2017 Day 18: Duet part 1 (need long as data type of register values ...)

+++ 2017/src/main/java/de/adventofcode/chrisgw/day18/RegisterReference.java
@@ -12,9 +12,9 @@ public class RegisterReference {
     }
 
 
-    public int getValue(Duet duet) {
+    public long getValue(Duet duet) {
         if (isStaticValue()) {
-            return Integer.parseInt(registerReference);
+            return Long.parseLong(registerReference);
         } else {
             return duet.getRegisterValue(getRegisterName());
         }
